Move RealmUtils to internal package
